requirements.txt
streamlit==1.29.0
numpy==1.26.4
Pillow==10.2.0
--force-reinstall
opencv-python==4.8.1.78
